+++ R/g_forest.R
@@ -28,7 +28,8 @@
 #'  column index from `tbl` containing data to be used to determine relative
 #'  size for estimator plot symbol. Typically, the symbol size is proportional to the
 #'  sample size used to calculate the estimator. If `NULL`, the same symbol
-#'  size is used for all subgroups.
+#'  size is used for all subgroups. By default tries to get this from
+#'  `tbl` attribute `col_symbol_size`, otherwise needs to be manually specified.
 #'
 #' @import grid
 #'
@@ -82,7 +83,49 @@
 #'
 #' draw_grob(p)
 #'
-#' # Works with any rtable
+#' # Survival forest plot example.
+#'
+#' adtte <- synthetic_cdisc_data("latest")$adtte
+#'
+#' # Save variable labels before data processing steps.
+#' adtte_labels <- var_labels(adtte)
+#'
+#' adtte_f <- adtte %>%
+#'   filter(
+#'     PARAMCD == "OS",
+#'     ARM %in% c("B: Placebo", "A: Drug X"),
+#'     SEX %in% c("M", "F")
+#'   ) %>%
+#'   mutate(
+#'     # Reorder levels of ARM to display reference arm before treatment arm.
+#'     ARM = droplevels(fct_relevel(ARM, "B: Placebo")),
+#'     SEX = droplevels(SEX),
+#'     AVALU = as.character(AVALU),
+#'     is_event = CNSR == 0
+#'   ) %>%
+#'   var_relabel(
+#'     ARM = adtte_labels["ARM"],
+#'     SEX = adtte_labels["SEX"],
+#'     AVALU = adtte_labels["AVALU"],
+#'     is_event = "Event Flag"
+#'   )
+#'
+#' df <- extract_survival_subgroups(
+#'   variables = list(
+#'     tte = "AVAL",
+#'     is_event = "is_event",
+#'     arm = "ARM", subgroups = c("SEX", "BMRKR2")
+#'   ),
+#'   data = adtte_f
+#' )
+#'
+#' table_hr <- basic_table() %>%
+#'   tabulate_survival_subgroups(df, time_unit = adtte_f$AVALU[1])
+#'
+#' g_forest(table_hr)
+#'
+#'
+#' # Works with any rtable.
 #'
 #' tbl <- rtable(
 #'   header = c("E", "CI", "N"),
@@ -129,7 +172,7 @@ g_forest <- function(tbl,
                      width_row_names = NULL,
                      width_columns = NULL,
                      width_forest = unit(1, "null"),
-                     col_symbol_size = 1,
+                     col_symbol_size = attr(tbl, "col_symbol_size"),
                      draw = TRUE,
                      newpage = TRUE) {