+++ tests/testthat/test-survival_duration_subgroups.R
@@ -57,6 +57,7 @@ test_that("extract_survival_subgroups functions as expected with valid input and
     hr = data.frame(
       arm = rep(" ", 6),
       n_tot = c(268, 161, 107, 95, 93, 80),
+      n_tot_events = c(166, 95, 71, 61, 55, 50),
       hr = c(
         0.71736505115489, 0.697969331159471, 0.783616674201674, 0.705072968604656,
         0.572806884078014, 0.976900177598777
@@ -142,6 +143,7 @@ test_that("extract_survival_subgroups functions as expected with NULL subgroups"
     hr = data.frame(
       arm = " ",
       n_tot = 268,
+      n_tot_events = 166,
       hr = 0.71736505115489,
       lcl = 0.527523110746632,
       ucl = 0.975526201857014,
@@ -157,7 +159,6 @@ test_that("extract_survival_subgroups functions as expected with NULL subgroups"
   )
 
   expect_equal(result, expected, tol = 0.000001)
-
 })
 
 test_that("a_survival_subgroups functions as expected with valid input", {
@@ -184,7 +185,6 @@ test_that("a_survival_subgroups functions as expected with valid input", {
     .Dim = c(3L, 3L)
   )
   expect_equal(result_matrix, expected_matrix)
-
 })
 
 test_that("tabulate_survival_subgroups functions as expected with valid input", {
@@ -203,23 +203,21 @@ test_that("tabulate_survival_subgroups functions as expected with valid input", 
   result_matrix <- to_string_matrix(result)
 
   expected_matrix <- structure(
-    c(
-      "Baseline Risk Factors", "", "All Patients", "Sex",
+    c("Baseline Risk Factors", "", "All Patients", "Sex",
       "F", "M", "Categorical Level Biomarker 2", "LOW", "MEDIUM", "HIGH",
-      " ", "Total n", "268", "", "161", "107", "", "95", "93", "80",
-      "B: Placebo", "n", "134", "", "82", "52", "", "45", "56", "33",
-      "B: Placebo", "Median (DAYS)", "837.4", "", "850.9", "527.7",
-      "", "751.4", "722.8", "848.2", "A: Drug X", "n", "134", "", "79",
-      "55", "", "50", "37", "47", "A: Drug X", "Median (DAYS)", "1260.5",
-      "", "1274.8", "849.3", "", "1160.6", "1269.4", "1070.8", " ",
-      "Hazard Ratio", "0.72", "", "0.70", "0.78", "", "0.71", "0.57",
-      "0.98", " ", "95% Wald CI", "(0.53, 0.98)", "", "(0.46, 1.05)",
+      " ", "Total Events", "166", "", "95", "71", "", "61", "55", "50",
+      "B: Placebo", "Events", "87", "", "50", "37", "", "30", "36",
+      "21", "B: Placebo", "Median (DAYS)", "837.4", "", "850.9", "527.7",
+      "", "751.4", "722.8", "848.2", "A: Drug X", "Events", "79", "",
+      "45", "34", "", "31", "19", "29", "A: Drug X", "Median (DAYS)",
+      "1260.5", "", "1274.8", "849.3", "", "1160.6", "1269.4", "1070.8",
+      " ", "Hazard Ratio", "0.72", "", "0.70", "0.78", "", "0.71",
+      "0.57", "0.98", " ", "95% Wald CI", "(0.53, 0.98)", "", "(0.46, 1.05)",
       "(0.49, 1.26)", "", "(0.42, 1.17)", "(0.32, 1.01)", "(0.56, 1.72)"
     ),
     .Dim = c(10L, 8L)
   )
   expect_equal(result_matrix, expected_matrix)
-
 })
 
 test_that("tabulate_survival_subgroups functions as expected with NULL subgroups", {
@@ -240,9 +238,10 @@ test_that("tabulate_survival_subgroups functions as expected with NULL subgroups
   expected_matrix <- structure(
     c(
       "Baseline Risk Factors", "", "All Patients", " ",
-      "Total n", "268", "B: Placebo", "n", "134", "B: Placebo", "Median (DAYS)",
-      "837.4", "A: Drug X", "n", "134", "A: Drug X", "Median (DAYS)",
-      "1260.5", " ", "Hazard Ratio", "0.72", " ", "95% Wald CI", "(0.53, 0.98)"
+      "Total Events", "166", "B: Placebo", "Events", "87", "B: Placebo",
+      "Median (DAYS)", "837.4", "A: Drug X", "Events", "79", "A: Drug X",
+      "Median (DAYS)", "1260.5", " ", "Hazard Ratio", "0.72", " ",
+      "95% Wald CI", "(0.53, 0.98)"
     ),
     .Dim = c(3L, 8L)
   )
@@ -270,19 +269,19 @@ test_that("tabulate_survival_subgroups functions as expected with extreme values
   expected_matrix <- structure(
     c(
       "Baseline Risk Factors", "", "All Patients", "Geographic Region 1",
-      "Asia", "North America", "South America", " ", "Total n", "30",
-      "", "5", "5", "20", "B: Placebo", "n", "11", "", "2", "2", "7",
-      "B: Placebo", "Median (DAYS)", "742.2", "", "176.4", "NA", "837.4",
-      "A: Drug X", "n", "19", "", "3", "3", "13", "A: Drug X", "Median (DAYS)",
-      "985.4", "", "985.4", "242.7", "1759.9", " ", "Hazard Ratio",
-      "0.92", "", "<0.01", ">999.99", "1.10", " ", "95% Wald CI", "(0.31, 2.76)",
-      "", "(0.00, >999.99)", "(0.00, >999.99)", "(0.27, 4.43)"
+      "Asia", "North America", "South America", " ", "Total Events",
+      "18", "", "4", "2", "12", "B: Placebo", "Events", "5", "", "2",
+      "0", "3", "B: Placebo", "Median (DAYS)", "742.2", "", "176.4",
+      "NA", "837.4", "A: Drug X", "Events", "13", "", "2", "2", "9",
+      "A: Drug X", "Median (DAYS)", "985.4", "", "985.4", "242.7",
+      "1759.9", " ", "Hazard Ratio", "0.92", "", "<0.01", ">999.99",
+      "1.10", " ", "95% Wald CI", "(0.31, 2.76)", "", "(0.00, >999.99)",
+      "(0.00, >999.99)", "(0.27, 4.43)"
     ),
     .Dim = 7:8
   )
 
   expect_equal(result_matrix, expected_matrix)
-
 })
 
 test_that("tabulate_survival_subgroups functions as expected when one arm has 0 records", {
@@ -298,7 +297,7 @@ test_that("tabulate_survival_subgroups functions as expected when one arm has 0 
   result <- basic_table() %>%
     tabulate_survival_subgroups(
       df = df,
-      vars = c("n_tot", "n", "n_events", "median", "hr", "ci", "pval"),
+      vars = c("n_tot_events", "n", "n_events", "median", "hr", "ci", "pval"),
       time_unit = adtte$AVALU[1]
     )
 
@@ -309,29 +308,71 @@ test_that("tabulate_survival_subgroups functions as expected when one arm has 0 
       "Baseline Risk Factors", "", "All Patients", "Race",
       "ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE", "AMERICAN INDIAN OR ALASKA NATIVE",
       "MULTIPLE", "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER", " ",
-      "Total n", "268", "", "135", "59", "53", "19", "1", "1", "B: Placebo",
-      "n", "134", "", "67", "28", "26", "11", "1", "1", "B: Placebo",
-      "Events", "87", "", "40", "20", "19", "6", "1", "1", "B: Placebo",
-      "Median (DAYS)", "837.4", "", "906", "751.4", "841.2", "741.9",
-      "33.6", "153.5", "A: Drug X", "n", "134", "", "68", "31", "27",
-      "8", "0", "0", "A: Drug X", "Events", "79", "", "42", "15", "16",
-      "6", "NA", "NA", "A: Drug X", "Median (DAYS)", "1260.5", "",
-      "1274.8", "1327.8", "774.7", "849.3", "NA", "NA", " ", "Hazard Ratio",
-      "0.72", "", "0.75", "0.63", "0.69", "1.01", "NA", "NA", " ",
-      "95% Wald CI", "(0.53, 0.98)", "", "(0.48, 1.17)", "(0.32, 1.25)",
-      "(0.35, 1.35)", "(0.28, 3.61)", "(NA, NA)", "(NA, NA)", " ",
-      "p-value (log-rank)", "0.0334", "", "0.2012", "0.1832", "0.2745",
-      "0.9839", "NA", "NA"
+      "Total Events", "166", "", "82", "35", "35", "12", "1", "1",
+      "B: Placebo", "n", "134", "", "67", "28", "26", "11", "1", "1",
+      "B: Placebo", "Events", "87", "", "40", "20", "19", "6", "1",
+      "1", "B: Placebo", "Median (DAYS)", "837.4", "", "906", "751.4",
+      "841.2", "741.9", "33.6", "153.5", "A: Drug X", "n", "134", "",
+      "68", "31", "27", "8", "0", "0", "A: Drug X", "Events", "79",
+      "", "42", "15", "16", "6", "NA", "NA", "A: Drug X", "Median (DAYS)",
+      "1260.5", "", "1274.8", "1327.8", "774.7", "849.3", "NA", "NA",
+      " ", "Hazard Ratio", "0.72", "", "0.75", "0.63", "0.69", "1.01",
+      "NA", "NA", " ", "95% Wald CI", "(0.53, 0.98)", "", "(0.48, 1.17)",
+      "(0.32, 1.25)", "(0.35, 1.35)", "(0.28, 3.61)", "(NA, NA)", "(NA, NA)",
+      " ", "p-value (log-rank)", "0.0334", "", "0.2012", "0.1832",
+      "0.2745", "0.9839", "NA", "NA"
     ),
     .Dim = 10:11
   )
   expect_equal(result_matrix, expected_matrix)
+})
 
+test_that("tabulate_survival_subgroups works correctly with both `n_tot` and `n_tot_events` in `vars`", {
+
+  adtte <- adtte %>%
+    preprocess_adtte()
+
+  df <- extract_survival_subgroups(
+    variables = list(tte = "AVAL", is_event = "is_event", arm = "ARM", subgroups = "RACE"),
+    data = adtte
+  )
+
+  # Both n_tot variables, but no surv time vars.
+  result_both <- basic_table() %>%
+    tabulate_survival_subgroups(
+      df = df,
+      vars = c("hr", "ci", "n_tot", "pval", "n_tot_events"),
+      time_unit = adtte$AVALU[1]
+    )
+  # Check that the column indices attributes are correct.
+  expected_cols_both <- list(col_x = 1L, col_ci = 2L, col_symbol_size = 3L)
+  result_cols_both <- attributes(result_both)[c("col_x", "col_ci", "col_symbol_size")]
+  expect_identical(result_cols_both, expected_cols_both)
+
+  # Both n_tot variables and also surv time vars, so we have a reordering of the vars in the table.
+  result_both_survtime <- basic_table() %>%
+    tabulate_survival_subgroups(
+      df = df,
+      vars = c("hr", "median", "n_events", "ci", "pval", "n_tot_events", "n", "n_tot"),
+      time_unit = adtte$AVALU[1]
+    )
+  # Check that the column indices attributes are correct.
+  expected_cols_both_survtime <- list(col_x = 9L, col_ci = 10L, col_symbol_size = 1L)
+  result_cols_both_survtime <- attributes(result_both_survtime)[c("col_x", "col_ci", "col_symbol_size")]
+  expect_identical(result_cols_both_survtime, expected_cols_both_survtime)
+  # Check header of table.
+  result_header_both_survtime <- to_string_matrix(result_both_survtime)[2, ]
+  expected_header_both_survtime <- c(
+    "", "Total Events", "Total n", "Median (DAYS)", "Events", "n",
+    "Median (DAYS)", "Events", "n", "Hazard Ratio", "95% Wald CI",
+    "p-value (log-rank)"
+  )
+  expect_identical(result_header_both_survtime, expected_header_both_survtime)
 })
 
 test_that("d_survival_subgroups_colvars functions as expected with valid input", {
 
-  vars <- c("n", "n_events", "median", "n_tot", "hr", "ci", "pval")
+  vars <- c("n", "n_events", "median", "n_tot_events", "hr", "ci", "pval")
 
   result <- d_survival_subgroups_colvars(
     vars = vars,
@@ -341,12 +382,12 @@ test_that("d_survival_subgroups_colvars functions as expected with valid input",
   )
 
   expected <- list(
-    vars = c("n", "n_events", "median", "n_tot", "hr", "lcl", "pval"),
+    vars = c("n", "n_events", "median", "n_tot_events", "hr", "lcl", "pval"),
     labels = c(
       n = "n",
       n_events = "Events",
       median = "Median (Months)",
-      n_tot = "Total n",
+      n_tot_events = "Total Events",
       hr = "Hazard Ratio",
       ci = "90% Wald CI",
       pval = "p-value (log-rank)"
@@ -354,5 +395,4 @@ test_that("d_survival_subgroups_colvars functions as expected with valid input",
   )
 
   expect_equal(result, expected)
-
 })