+++ NEWS.md
@@ -10,6 +10,7 @@ with marked laboratory abnormalities are counted.
 * Updated `cox_regression` to be able to work without covariates. 
 * Delete `count_abnormal_by_worst_grade_by_baseline` and its related statistic and analysis function as `LBT13` and 
 `LBT14` are adopting new design.
+* Corrected `tabulate_survival_subgroups` and related survival forest plot functions to use total number of events, instead of observations, as default for scaling the symbol sizes in the plot. (The user might still use total number of observations manually if they wish to do so.)
 
 # tern 0.7.3